% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss.R
\name{diss}
\alias{diss}
\title{Density Inclusive Study Size (DISS) metric}
\usage{
diss(x, c)
}
\arguments{
\item{x}{the running variable.}

\item{c}{the cutoff.}
}
\value{
a vector of the following values:
\itemize{
\item \code{m} is the number of observations within the \code{h.diss} bandwidth of the cutoff.
\item \code{m.l} is the number of observations within the \code{h.diss} bandwidth below the cutoff.
\item \code{m.r} is the number of observations within the \code{h.diss} bandwidth above the cutoff.
\item \code{h.diss} is the rule of thumb bandwidth
}
}
\description{
Calculates the number of observations within a Silverman rule of thumb bandwidth of the cutoff.
}
