% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiana_school_scores}
\alias{indiana_school_scores}
\title{Indiana School Scores Data}
\format{
\subsection{\code{indiana_school_scores}}{

A data frame with 1,933 rows and 2 columns:
\describe{
\item{score2017}{The school's overall points score in 2017}
\item{score2018}{The school's overall points score in 2018}
...
}
}
}
\source{
\url{https://www.in.gov/doe/files/af-2017-school-grade-placement-03192018.xlsx}

\url{https://www.in.gov/doe/files/2018-state-school-grade-results.xlsx}
}
\usage{
indiana_school_scores
}
\description{
A subset of 2017 and 2018 school scores from the Indiana Department of Education
}
\keyword{datasets}
