% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdple.R
\name{rdple}
\alias{rdple}
\title{Partial linear estimator for regression discontinuity designs}
\usage{
rdple(
  y,
  x,
  c,
  h = "ska",
  degree = 1,
  kernel = "epanechnikov",
  sparse.adj = FALSE,
  sparse.r = 1
)
}
\arguments{
\item{y}{the response variable.}

\item{x}{the running variable.}

\item{c}{the cutoff.}

\item{h}{the bandwidth to be used (if not provided, a bandwidth will be calculate using ska_band.}

\item{degree}{the degree of the local polynomial weights (can be 0 or 1).}

\item{kernel}{specifies the kernel function used for the local polynomial weights. Options are \code{epanechnikov} (default) or \code{triangular}.}

\item{sparse.adj}{whether or not to use the sparsity adjustment for the SKA bandwidth, can be \code{FALSE} (default) or \code{TRUE}.}

\item{sparse.r}{if \code{sparse.r = TRUE}, this is the positive number of points that are required to reach across the cutoff.}
}
\value{
A list containing the effect estimate, the standard error of the effect estimate, and the bandwidth.
}
\description{
Implements the partial linear estimator for regression discontinuity designs.
If a bandwidth is not provided, uses ska_band to calculate the Swartzentruber-KAizar bandwidth.
}
