% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ska_band.R
\name{ska_band}
\alias{ska_band}
\title{Title}
\usage{
ska_band(y, x, c, kernel = "epanechnikov")
}
\arguments{
\item{y}{A numeric vector.}

\item{x}{A numeric vector.}

\item{c}{A number.}

\item{kernel}{A character string.}
}
\value{
A number.
}
\description{
Title
}
