% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdple.R
\name{rdple}
\alias{rdple}
\title{Title}
\usage{
rdple(
  y,
  x,
  c,
  h = "ska",
  degree = 1,
  kernel = "epanechnikov",
  sparse.adj = FALSE,
  sparse.r = 1
)
}
\arguments{
\item{y}{A numeric vector.}

\item{x}{A numeric vector.}

\item{c}{A number.}

\item{h}{A number or a character string.}

\item{degree}{A number.}

\item{kernel}{A character string.}

\item{sparse.adj}{A character string.}

\item{sparse.r}{A number.}
}
\value{
A list.
}
\description{
Title
}
