% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ska_band.R
\name{ska_band}
\alias{ska_band}
\title{Swartzentruber-KAizar bandwidth algorithm}
\usage{
ska_band(y, x, c, kernel = "epanechnikov")
}
\arguments{
\item{y}{the response variable.}

\item{x}{the running variable.}

\item{c}{the cutoff.}

\item{kernel}{specifies the kernel function used for the local polynomial weights. Options are \code{epanechnikov} (default) or \code{triangular}.}
}
\value{
The SKA bandwidth.
}
\description{
Calculates the Swartzentruber-KAizar (SKA) bandwidth value for an RD design.
}
